# backend/openapi.yaml
openapi: 3.0.3
info:
  title: NFT Indexer API
  version: 1.0.0
  description: API for querying indexed NFT data from Starknet contracts.
servers:
  - url: http://localhost:3000/api
    description: Local development
paths:
  /assets:
    get:
      summary: Get all assets
      parameters:
        - in: query
          name: indexerSource
          schema: { type: string, enum: [MEDIALANO-DAPP, MEDIALANO-MIPP] }
        - in: query
          name: collectionId
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
        - in: query
          name: sortBy
          schema:
            { type: string, enum: [mintedAtBlock, id], default: mintedAtBlock }
        - in: query
          name: sortOrder
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAssets"
...
  /collections:
    get:
      summary: Get all collections
      parameters:
        - in: query
          name: indexerSource
          schema: { type: string }
        - in: query
          name: creator
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCollections"
...
  /transfers:
    get:
      summary: Get all transfers
      parameters:
        - in: query
          name: tokenId
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string }
        - in: query
          name: to
          schema: { type: string }
        - in: query
          name: indexerSource
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
        - in: query
          name: sortOrder
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        "200":
          description: Transfer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTransfers"
...
components:
  schemas:
    Asset:
      type: object
      properties:
        id: { type: string }
        collectionId: { type: string }
        owner: { type: string }
        tokenUri: { type: string }
        mintedAtBlock: { type: integer }
        indexerSource: { type: string }
    Collection:
      type: object
      properties:
        id: { type: string }
        creator: { type: string }
        metadataUri: { type: string }
        createdAtBlock: { type: integer }
        indexerSource: { type: string }
        assetCount: { type: integer }
    Transfer:
      type: object
      properties:
        id: { type: string }
        tokenId: { type: string }
        from: { type: string }
        to: { type: string }
        block: { type: integer }
        indexerSource: { type: string }
...
    PaginatedAssets:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Asset" }
        pagination:
          type: object
          properties:
            total: { type: integer }
            limit: { type: integer }
            offset: { type: integer }
            hasMore: { type: boolean }
    PaginatedCollections:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Collection" }
        pagination:
          type: object
          properties:
            total: { type: integer }
            limit: { type: integer }
            offset: { type: integer }
            hasMore: { type: boolean }
    PaginatedTransfers:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Transfer" }
        pagination:
          type: object
          properties:
            total: { type: integer }
            limit: { type: integer }
            offset: { type: integer }
            hasMore: { type: boolean }